o Using Get/SetByteArrayRegion instead of Get/SetByteArrayElements/ReleaseByteArrayElements.
This significantly improves performance when dealing with large arrays as it prevents one extra copy operation of the array. The downside is 64k of extra stack usage. Reads are limited to this 64k buffer instead of copying everything in one go. When changing from the current implementation to this version we saw an 6x performance gain for 10MB arrays.

This follows the same pattern as OpenJDK 8 uses:
http://hg.openjdk.java.net/jdk8/jdk8/jdk/file/687fd7c7986d/src/solaris/native/java/net/SocketInputStream.c
http://hg.openjdk.java.net/jdk8/jdk8/jdk/file/687fd7c7986d/src/solaris/native/java/net/SocketOutputStream.c