Support building with and without linking against libc
The trick to not link against libc (in order to support musl-libc based
distributions such as Alpine Linux) doesn't work for all Linux
distributions.

We're now trying to load the properly-linked native library first, and
upon an UnsatisfiedLinkError, we try the version without declared
dependencies ("nodeps").

To simplify building, we introduce a custom flag to crossclang,
-Xcrossclang-with-and-without-lc, which will, upon linking, create two
libraries instead of one.