Update Xcode project to use crossclang toolchain
This lets us use Xcode to compile for all supported architectures at
once. Specifically, we get all the warnings from all platforms.

In order to use this Xcode project, you need to install the provided
crossclang.xctoolchain and crossclang.sdk, and switch to the crossclang
toolchain in Xcode. Of course, you also need to have all targets in
/opt/crossclang/target-sdks.