Fix code coverage, multi-release jar building
jacoco doesn't support multi-release jars yet, and previously, we were
modifying the classes in outputDirectory in a way that made it
impossible to get an accurate aggregate code coverage report.

We now build the multi-release jar outside of the target/classes
directory, and add a "junixsocket-codecoverage" module that can
correctly detect coverage across sub-modules.