Remove AFUNIXSocketImpl#finalize for Java 9+
We can now get rid of #finalize for Java 9 and later. This change
introduces CleanableState, a class with two different implementations,
one for Java 9 and newer, and one for Java 8 and older, which hides the
finalize/Cleaner logic under the hood.

This change also introduces a test case for the issue mentioned in
https://github.com/kohlschutter/junixsocket/pull/29