AFUNIXSocketCredentials: add #isEmpty
Add a method that checks if any credentials were actually set (which may
not be the case despite the corresponding capability being set).

For example, this may happen currently on NetBSD with datagram sockets.

This change amends the PeerCredentialsTest accordingly, so a warning
message is emitted instead of an exception.