Allow cross-compiling for NetBSD
Before this change, cross-compilation suceeded, but the resulting shared
library would not load ("wrong number of segments (4 != 2). This cryptic
error message refers to the number of PT_LOAD segments that clang lld
created (4) and the number that NetBSD expects (exactly 2), which can
be inspected by (llvm-)readelf --segments library.so.

There seems to be some conflict between clang/llvm and NetBSD
maintainers about how to properly link with NetBSD as a target [1], and
it doesn't look like this is going to be fixed soon.

Going through the proposed (and presumably stale) LLVM patch to enable a
NetBSD-specific linker [2], it seems that all it takes to get our
library to build for NetBSD is to add a few linker options. It turns out
one more needs to be added, otherwise we end up with 3 segments.

Here's the set of flags that we use:

-Wl,--no-rosegment,--disable-new-dtags,-znognustack,-znoseparate-code,-znorelro

[1] https://blog.netbsd.org/tnf/entry/the_first_report_on_lld
[2] https://reviews.llvm.org/D70048