Handle access denied condition on /dev/vsock, improve tests
With AF_VSOCK, older Linux kernels may have problems accessing sockets
that were asked to have a LOCAL (1) CID, which we resolved to HOST (2),
when the corresponding /dev/vsock device is not readable by the process.

Introduce a new set of SocketExceptions (InvalidSocketException and its
subclasses), and change the unit test code to handle these conditions
properly.