native: win32: Terminate accept call on rebound AF_UNIX socket
On Windows, when rebinding an existing socket address, a previous accept
call needs to be interrupted manually.

We do this by renaming the socket file to a temporary path in the same
directory and connect to it. Upon accept, we always check if the bound
file name still exists. If it doesn't, we shut down the server socket
and throw an exception.

https://github.com/kohlschutter/junixsocket/issues/121