Fix AcceptTimeoutTest: Terminate helper thread, use correct address
When connecting to a server, the server address should not be retrieved
after the server socket is shut down, since this may return a wildcard
address (instead of an invalid one), for example on AF_VSOCK, resulting
in false-positive test failures.

Moreover, we need to terminate the helper thread upon timeout.