AFCore: Fix potential hang in send
When using non-direct ByteBuffers (the default), sending data to a
socket may cause an infinite loop.

This is caused by an incorrect position calculation in case the position
of the ByteBuffer that is to be written is different from the position
of the ThreadLocal-provided ByteBuffer (usually when the position != 0).

This in turn may cause the native code around send(2) to erroneously cap
the number of bytes to be sent to 0.

Fix the send logic in AFCore to separate position in the source buffer
and the ThreadLocal-provided direct ByteBuffer.