native: Add code support for Android
The Android Java SDK contains some changes that need to be addressed:

- FileDescriptor int access is available via helper methods instead of a
  field
- AbstractSelectableChannel: "removeKey" is called "deregister"
- ServerSocket: We cannot override "impl" (unless "hidden_api_policy" is
  enabled).

Moreover, we used to call CallObjectMethodA instead of CallVoidMethodA
for a void method.  Android's JNI is rightfully picky about this, so
let's fix that as well.

Lastly, improve error reporting for "handleFieldNotFound" by reporting
the full classname, just just the simple name.