Fix non-direct ByteBuffer write
We use a cached direct ByteBuffer upon attempting to write a non-direct
ByteBuffer.

However, there was a logic bug that messed with the limit of the source
buffer in an incorrect way.  This manifested in AFSocketServerConnector
sending the wrong data for response buffers of sizes larger than the
configured direct byte buffer size.

Replace the limit-logic with a "hasRemaining" check and loop the put()
as long as both buffers have something to send/receive.
Add a test case for the reported bug with AFSocketServerConnector.
Also add some precautious logic for the read case, too.

https://github.com/kohlschutter/junixsocket/issues/135