native: Fix a hang upon send on macOS/BSD caused by ENOBUFS
On macOS (and supposedly some BSDs), send/sendto may not block when
sending datagrams. This may result in buffer depletion / ENOBUFS upon
send.

The workaround is to call sched_yield and hope for the best. Previously
however, we were erroneously not calling sched_yield on systems that
supported it. We were trying to poll for available buffer space.

That in turn however, did not work on macOS.

Fix the sched_yield workaround, and use it on all systems (use
SwitchToThread on Win32).