Fix reading into a non-direct ByteBuffer with position() != 0
Previously, we were wrongly reusing the destination ByteBuffer
position for our cached direct-ByteBuffer, which usually has a position
of 0.

This had the undesired effect that bytes were ending up in a wrong
position.

Separate the two positions. Also restore the destination buffer limit,
which we may have changed during put().