Embed git properties in code; improve AFSocket.getVersion, selftest
By embedding the actual git commit information and project version into
the Java code, we can make debugging a lot easier, especially in
environments that don't keep the META-INF files around that we relied
upon earlier (e.g., Android).

This relies on latest changes in kohlschutter-parent 1.5.5-SNAPSHOT,
which is required for building junixsocket.