NativeLibraryLoader: Try loading from user.home, user.dir if tmp fails
Unless specified otherwise, junixsocket tries extracting the native
library from the classpath, storing it as a temporary file before it is
loaded with System.load.

In some environments, /tmp may be unsuitable for loading that temporary
file as a native library, for example because the mountpoint is mounted
with the "noexec" option.

If loading fails from the temporary directory due to permissions, try
loading from user.home or user.dir, instead.