darwin: Add junixsocket-darwin, with initial support for AF_SYSTEM
Add the "junixsocket-darwin" module.

Presently, this adds support for the AF_SYSTEM domain, which is used for
communication with the Darwin kernel.

Due to the nature of this socket domain, custom tests are provided. One,
UtunTest, is actually pretty cool but it requires root privileges to
run. It establishes a VPN ("user tunnel"), pings the destination, parses
the ping packet and answers with a hand-crafted ICMP echo response.

No support for SYSPROTO_EVENT yet.