pom: Guard native-custom dependency with junixsocket.native-custom.skip
When working with SNAPSHOT builds, junixsocket-native-custom may not be
available as a dependency. It is marked as an "optional" dependency for
testing, however building still fails if the artifact is not available.

Allow users to disable inclusion of the junixsocket-native-custom
dependency by specifying -Djunixsocket.native-custom.skip when running
Maven:

mvn clean install -Djunixsocket.native-custom.skip \
  -rf :junixsocket-common