AFSocketAddress: Simplify creation of unresolved InetAddress
Previously, we were relying on deserialization to craft an Inet4Address
that has both a non-resolvable hostname and a local IP address, which we
use to identify junixsocket-specific addresses.

The code we relied on is different on Android, causing these addresses
to be looked up using the Android DNS resolver, which is guaranteed to
fail, just waiting time.

Use InetAddress.getByAddress(String hostname, byte[4] ipv4Addr) which is
specified to not use DNS resolution.

https://github.com/kohlschutter/junixsocket/issues/139