AFSocketAddress: Prevent false-positive "mutable object" warnings
When passing AFSocketAddress as constructor arguments, spotbugs may
erroneously claim that we expose class internals due to the address
being mutable.

In fact, all mutable fields of AFSocketAddress are only mutable so we
can support deserialization, and are not modifiable otherwise.

Add @Immutable, and suppress some corresponding warnings in code.