NativeLibraryLoader: Don't check UnsatisfiedLinkError message
We currently check the message text of an UnsatisfiedLinkError to decide
if we should try to load a library from user.home/user.dir when loading
from the temporary directory fails. The idea is that only certain
failures merit that fallback.

Unfortunately, checking the message text of an UnsatisfiedLinkError is
not reliable.

Since this operation only happens once upon load, remove that
optimization, and try all other paths regardless of message text.

https://github.com/kohlschutter/junixsocket/issues/140