AFSelector: Use single ConcurrentHashMap for registered/selected keys
Since we already have a ConcurrentHashMap, let's make use of the value,
which now indicates the "selected" state.

Introduce MapValueSet, which is a view over elements of the
"keysRegistered" map, and precisely only those elements that have a
certain value.

For each call to select, we increment the expected value, and then set
only the actually selected entries to that value, so we don't have to
clear the entire map.

https://github.com/kohlschutter/junixsocket/issues/145