Address Spotbugs error "CT_CONSTRUCTOR_THROW"
A new Spotbugs check unconvered that several classes were affected by
potential "Finalizer attack", which would allow the retrieval of
half-initialized classes by a nefarious subclass.

See
https://wiki.sei.cmu.edu/confluence/display/java/OBJ11-J.+Be+wary+of+letting+constructors+throw+exceptions
for details.

The impact of this bug appears to be small, but the issue should be
fixed nevertheless.

In many cases, a declared exception was actually not needed, so it was
removed. In other cases, the class was unnecessarily non-final, so they
were marked final. Then, for some cases, adding an empty, final
finalizer method was deemed acceptable.

Lastly, for some classes, no good solution other than ignoring the
warning could be established. In these cases, it was deemed necessary
to leave default finalization on, and the risk was considered
negligible. Should a better solution be found, these remaining classes
will improved later.

The fallout of this change is considered minimal, as subclassing one of
these affected classes by users of junixsocket is not expected.