common: FileDescriptorCast: Update blocking state as necessary
Previously, casting a FileDescriptor to Socket would not reset its
(cached) blocking state to "blocking" if it was configured non-blocking,
which may lead to unexpected behavior.

Check the "blocking" state upon casting and set it accordingly when a
Socket/DatagramSocket/ServerSocket is required (since they are expected
to be blocking by default).

When casting to a SocketChannel/DatagramChannel/ServerSocketChannel,
adjust the Java-cached "blocking" state to the actual blocking state of
the native socket.

Moreover, add support to declare "we don't know the state" (like in
Windows, which is currently not supported for casting, though), and
ensure we get the cached state and the native state back in sync.

https://github.com/kohlschutter/junixsocket/issues/151