common: Fix NativeLibraryLoader for AIX/IBM i on Java 15 and newer
With OpenJ9 15 or newer on AIX and IBM i, System.mapLibraryName returns
".so" as the library suffix. Before that, it was ".a" on AIX and
".srvpgm".

This broke our native library loading on these platforms, as the
libraries stored in junixsocket-native-common were no longer found.

Add a check for the ".so" suffix and restore the expected suffixes.

Many thanks to Fabrice Bourquin for finding this issue!

https://github.com/eclipse-openj9/openj9/issues/9788