common,native: FileDescriptorCast: Add duplicating, clarify GC impact
When casting a FileDescriptor that belongs to a socket that is already
managed by junixsocket, Garbage Collection may close the underlying
resource if the original object is no longer reachable, which may happen
at an inopportune moment for the cast target.

Clarify why this happens and how this can be avoided (keep a reference
to the original socket around) or circumvented. Add ".duplicating"
(which calls "dup" under the hood). Update the corresponding tests.

Moreover, disable cleanup upon garbage collection when casting to
junixsocket socket implementations, as this is already handled
elsewhere.