Improve handling of static SelectorProvider open(ProtocolFamily) methods
Previously, some static methods that open SocketChannels etc. taking a
ProtocolFamily enum as a parameter were not overridden with AF* specific
code.

This may have lead to confusion as to which methods to use.

Provide proper implemetations, and also add a fallback SelectorProvider
for AFSelectorProvider.provider(), which forwards any requests to the
code responsible for the corresponding junixsocket ProtocolFamily.