test: Rework InterruptIssue158Test
Previously, the test may fail sporadically due to a race condition
between receiving a "thread interrupted" state on a live connection and
one that has been closed already (resulting in an
AsynchronousCloseException instead of a ClosedByInterruptException).

Make sure that the client connection is closed from the servce side only
after a delay that is significantly longer than any other delay in this
test class. Run that close() in a separate Thread to avoid unnecessarily
long delays.

Finally, set all expected exceptions back to ClosedByInterruptException,
as intended by the original author (cenodis).

https://github.com/kohlschutter/junixsocket/issues/158