native: Improve handling of "stat" variants
Depending on the exact setup, Java on Linux may not have access to the
symbol named "stat", and instead uses an __xstat function instead.
The availability of these functions vary across target systems.

Currently, we build with certain Linux SDKs and assume that they will
just work fine one way or another (i.e., by either linking with glibc or
not, via the nodeps variant).

Apparently, this doesn't work with Loongnix 20.6, which uses an older
glibc than what we expected from other loongarch64 systems we tested.

This is a good opportunity to rework the "stat" symbol lookup and
properly check for the availability of __xstat/stat by delaring them as
weak symbols and taking it from there.

https://github.com/kohlschutter/junixsocket/issues/161